requirements.txt
streamlit
openai
pillow
base64
